feat(consumer): start consumer once automatically.
